final_app/requirements.txt
annotated-types==0.7.0
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.23.11
awscrt==0.23.8
Babel==2.9.1
blinker==1.9.0
boto3==1.38.27
botocore==1.38.27
cachetools==5.5.2
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
click==8.1.8
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
cryptography==36.0.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.8
Flask==3.1.1
git-remote-codecommit==1.17
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gpg==1.15.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==2.10
ikp3db==1.4.2
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
libcomps==0.1.20
lockfile==0.12.2
MarkupSafe==3.0.2
mercurial==5.7.1
mysql-connector-python==9.3.0
netifaces==0.10.6
oauthlib==3.0.2
packaging==21.3
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.20
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.8.1
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
rsa==4.9.1
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
s3transfer==0.13.0
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
support-info==1.0
systemd-python==235
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==3.1.3
zipp==3.22.0
